+++ stable/fblbkd.h
@@ -1,5 +1,5 @@
 /*
-	Copyright (C) 1999 Claude SIMON (http://q37.info/contact/).
+	Copyright (C) 1999-2017 Claude SIMON (http://q37.info/contact/).
 
 	This file is part of the Epeios framework.
 
@@ -191,7 +191,7 @@ namespace fblbkd {
 		const char *Prefix_;
 		// Libelle du type de l'objet, du module.
 		const char *Name_;
-		// L'interface auquel le module est rattachï¿½.
+		// L'interface auquel le module est rattaché.
 		class backend___ *Backend_;
 	public:
 		//r The description of the request.
@@ -318,7 +318,7 @@ namespace fblbkd {
 		virtual sIndex FBLBKDNew( void )
 		{
 			qRFwk();
-			return 0;	// Pour ï¿½viter un warning.
+			return 0;	// Pour éviter un warning.
 		}
 		//v To delete the object with index 'Index'.
 		virtual void FBLBKDDelete( sIndex Index )
@@ -329,13 +329,13 @@ namespace fblbkd {
 		virtual void *FBLBKDObject( sIndex Index ) const
 		{
 			qRFwk();
-			return NULL;	// Pour ï¿½viter un 'warning'
+			return NULL;	// Pour éviter un 'warning'
 		}
 #if 0
 		//v To get the raw messages.
 		virtual void FBLBKDGetRawMessages( messages_ &Messages ) = 0;
 #endif
-		// Fonction appelï¿½e pour traiter la requï¿½te 'Requete' pour l'objet d'index 'Index'.
+		// Fonction appelée pour traiter la requête 'Requete' pour l'objet d'index 'Index'.
 		virtual void Handle_(
 			sIndex Index,
 			rRequest &Requete,
@@ -662,7 +662,7 @@ namespace fblbkd {
 		}
 	};
 
-	// Module maï¿½tre, qui fait tout le boulot.
+	// Module maître, qui fait tout le boulot.
 	class rMasterModule
 	: public rModule
 	{
@@ -748,7 +748,6 @@ namespace fblbkd {
 	class backend___
 	{
 	private:
-		bso::bool__ _CompatibilityTested;
 		fblovl::eMode Mode_;
 		TOL_CBUFFER___ _ClientOrigin;
 		TOL_CBUFFER___ _APIVersion;
@@ -756,14 +755,14 @@ namespace fblbkd {
 		TOL_CBUFFER___ _Language;
 		const lcl::locale_ *_Locale;
 		TOL_CBUFFER___ _BackendLabel;
-		// Informations ï¿½ propos du 'backend'.
+		// Informations à propos du 'backend'.
 		TOL_CBUFFER___ _ExtendedBackendInformations;
 		TOL_CBUFFER___ _BackendCopyright;
 		TOL_CBUFFER___ _SoftwareInformations;
 		str::wString
 			Key_,	// Key used for encrypting the codes.
 			Code_;	// Code to allow a blocking ping or a crash.
-		// Retourne le module correspondant ï¿½ 'IdType'.
+		// Retourne le module correspondant à 'IdType'.
 		rModule &Module_( type__ IdType )
 		{
 			if ( IdType != FBLBKD_MASTER_TYPE )
@@ -778,7 +777,7 @@ namespace fblbkd {
 			else
 				return Master_;	// Not very happy about this conversion, 
 		}
-		// Retourne le module correspondant ï¿½ 'IdObjet'.
+		// Retourne le module correspondant à 'IdObjet'.
 		rModule &Module_( object__ IdObjet )
 		{
 			return Module_( Type_( IdObjet ) );
@@ -787,12 +786,12 @@ namespace fblbkd {
 		{
 			return Module_( Type_( IdObjet ) );
 		}
-		// Retourne le type correpondant ï¿½ l'objet d'indetificateur 'IdObjet'.
+		// Retourne le type correpondant à l'objet d'indetificateur 'IdObjet'.
 		type__ Type_( object__ IdObjet ) const
 		{
 			return Links.Type( IdObjet );
 		}
-		// Retourne l'indexcorrespondant ï¿½ l'objet d'identificateur 'IdObjet'.
+		// Retourne l'indexcorrespondant à l'objet d'identificateur 'IdObjet'.
 		sIndex Index_( object__ IdObjet ) const
 		{
 			return Links.Index( IdObjet );
@@ -802,10 +801,6 @@ namespace fblbkd {
 			const char *APIVersion,
 			const char *MessageLabel,
 			const char *URLLabel );
-		bso::bool__ _TestCompatibility(	fdr::rRWDriver &FrontendIODriver );
-		bso::bool__ _HandleRequest(
-			fdr::rRWDriver &FrontendIODriver,
-			log_functions__ &LogFunctions );
 	protected:
 		virtual void *FBLBKDUserPointer( void ) = 0;
 	public:
@@ -815,13 +810,12 @@ namespace fblbkd {
 		links Links;
 		void reset( bso::bool__ P = true )
 		{
-			_CompatibilityTested = false;
 			Mode_ = fblovl::m_Undefined;
 			Key_.reset( P );
 			Code_.reset( P );
 		}
 		E_CVDTOR( backend___ );
-		// '[Backend|Publisher]Informations' ne sont PAS dupliquï¿½. Leur contenu de doit pas ï¿½tre modifiï¿½.
+		// '[Backend|Publisher]Informations' ne sont PAS dupliqué. Leur contenu de doit pas être modifié.
 		void Init(
 			fblovl::eMode Mode,
 			const char *APIVersion,
@@ -842,7 +836,6 @@ namespace fblbkd {
 			Modules.Init();
 			Links.Init();
 
-			_CompatibilityTested = false;
 			Mode_ = Mode;
 
 			str::string( BackendLabel ).Convert( _BackendLabel );
@@ -879,18 +872,10 @@ namespace fblbkd {
 			Module.Backend_ = this;
 			Modules.Append( &Module );
 		}
-		/*f Handle the request which come by 'Channel' and write the answer to 'Channel'.
-		If 'true' is returned, than the request contains a deconnection request. */
-		bso::bool__ Handle(
+		bso::bool__ TestCompatibility( fdr::rRWDriver &FrontendIODriver );
+		bso::bool__ HandleRequest(
 			fdr::rRWDriver &FrontendIODriver,
-			log_functions__ &LogFunctions )
-		{
-			if ( !_CompatibilityTested ) {
-				_CompatibilityTested = true;
-				return _TestCompatibility( FrontendIODriver );
-			}else
-				return _HandleRequest( FrontendIODriver, LogFunctions );
-		}
+			log_functions__ &LogFunctions );
 		/*f Return the command corresponding at request description 'Description' and
 		object type 'Type'. 'FBLBKD_INVALID_COMMAND' is returned if command not found. */
 		command__ Command(